services: Remove prefix and move into folder
The SQL migration is modified instead of added upon as it is a
backwards-compatible change (ie. servers.py can just ignore the
command_prefix field if it exists in the db).
