+++ sooch/services/servers.py
@@ -15,8 +15,7 @@ class Servers:
         """Return the server info associated with the ID requested"""
         cursor = self.database.connection.cursor()
         cursor.execute(
-            ("SELECT `name`, `command_prefix` FROM `server` "
-             "WHERE `discord_id` = ?"),
+            "SELECT `name` FROM `server` WHERE `discord_id` = ?",
             (discord_id,)
         )
         row = cursor.fetchone()
@@ -26,18 +25,15 @@ class Servers:
 
         return Server(
             discord_id,
-            row[0],
-            row[1]
+            row[0]
         )
 
     async def add_server(self, server: "Server"):
         """Add the provided server into the database"""
         cursor = self.database.connection.cursor()
         cursor.execute(
-            ("INSERT INTO `server`"
-             "(`discord_id`, `name`, `command_prefix`)"
-             "VALUES(?, ?, ?)"),
-            (server.discord_id, server.name, server.command_prefix)
+            ("INSERT INTO `server` (`discord_id`, `name`) VALUES(?, ?)"),
+            (server.discord_id, server.name)
         )
         self.database.connection.commit()
 
@@ -47,4 +43,3 @@ class Server:
     """Represent an instance of a guild in Discord."""
     discord_id: int
     name: str
-    command_prefix: str = "s!"
